Reorganized types registration code
