--- lib/registerTypes/TypesMapObjects3.cpp
@@ -1,32 +1,0 @@
-/*
- * TypesMapObjects3.cpp, part of VCMI engine
- *
- * Authors: listed in file AUTHORS in main folder
- *
- * License: GNU General Public License v2.0 or later
- * Full text of license available in license.txt file, in main folder
- *
- */
-#include "StdInc.h"
-#include "RegisterTypes.h"
-
-#include "../StartInfo.h"
-#include "../mapObjects/CObjectHandler.h"
-#include "../CCreatureHandler.h"
-#include "../VCMI_Lib.h"
-#include "../CArtHandler.h"
-#include "../CHeroHandler.h"
-#include "../spells/CSpellHandler.h"
-#include "../CTownHandler.h"
-
-#include "../serializer/BinaryDeserializer.h"
-#include "../serializer/BinarySerializer.h"
-#include "../serializer/CTypeList.h"
-
-VCMI_LIB_NAMESPACE_BEGIN
-
-template void registerTypesMapObjectTypes<BinaryDeserializer>(BinaryDeserializer & s);
-template void registerTypesMapObjectTypes<BinarySerializer>(BinarySerializer & s);
-template void registerTypesMapObjectTypes<CTypeList>(CTypeList & s);
-
-VCMI_LIB_NAMESPACE_END
