+++ lib/serializer/BinarySerializer.h
@@ -115,11 +115,7 @@ public:
 	bool smartPointerSerialization;
 	bool saving;
 
-	BinarySerializer(IBinaryWriter * w): CSaverBase(w)
-	{
-		saving=true;
-		smartPointerSerialization = true;
-	}
+	BinarySerializer(IBinaryWriter * w);
 
 	template<typename Base, typename Derived>
 	void registerType(const Base * b = nullptr, const Derived * d = nullptr)
@@ -401,32 +397,6 @@ public:
 			auto writ = static_cast<uint64_t>(data.to_ulong());
 			save(writ);
 		}
-	}
-};
-
-class DLL_LINKAGE CSaveFile : public IBinaryWriter
-{
-public:
-	BinarySerializer serializer;
-
-	boost::filesystem::path fName;
-	std::unique_ptr<std::fstream> sfile;
-
-	CSaveFile(const boost::filesystem::path &fname); //throws!
-	~CSaveFile();
-	int write(const void * data, unsigned size) override;
-
-	void openNextFile(const boost::filesystem::path &fname); //throws!
-	void clear();
-	void reportState(vstd::CLoggerBase * out) override;
-
-	void putMagicBytes(const std::string &text);
-
-	template<class T>
-	CSaveFile & operator<<(const T &t)
-	{
-		serializer & t;
-		return * this;
 	}
 };
